+++ lib/opencl_brute/worker/generic/pbkdf2_sha256_32.cl
@@ -65,7 +65,7 @@ typedef struct {
             /* Prints little endian, since that's what we use */   \
             v = (v >> r) & 0xFF;                \
             if (hex) {                          \
-                printf("%02x", v);              \
+                printf("%02x", (unsigned int)v); \
             } else {                            \
                 printf("%c", (char)v);          \
             }                                   \
